chore : 빌드 오류로 인한 폴더 구조 변경
