+++ src/refactoring/components/admin/CouponForm.tsx
@@ -1,34 +1,45 @@
-import { useAtomValue, useSetAtom } from "jotai";
-import { newCouponAtom } from "../../store/coupons/atom.ts";
-import { handleAddCouponAtom } from "../../store/coupons/action.ts";
-import CouponInput from "./CouponInput.tsx";
-import CouponSelect from "./CouponSelect.tsx";
+import { Coupon } from "@/types";
+import CouponInput from "./CouponInput";
+import CouponSelect from "./CouponSelect";
 
-export const CouponForm = () => {
-  const newCoupon = useAtomValue(newCouponAtom);
-  const handleAddCoupon = useSetAtom(handleAddCouponAtom);
+interface Props {
+  newCoupon: Coupon;
+  updateNewCoupon: (field: keyof Coupon, value: any) => void;
+  onAddCoupon: () => void;
+}
 
+const CouponForm = ({ newCoupon, updateNewCoupon, onAddCoupon }: Props) => {
   return (
     <div className="space-y-2 mb-4">
       <CouponInput
         type={"text"}
         value={newCoupon.name}
-        placeholder="쿠폰 이름"
-        field="name"
+        placeholder={"쿠폰 이름"}
+        updateNewCoupon={updateNewCoupon}
+        field={"name"}
       />
-
       <CouponInput
         type={"text"}
         value={newCoupon.code}
-        placeholder="쿠폰 코드"
-        field="code"
+        placeholder={"쿠폰 코드"}
+        updateNewCoupon={updateNewCoupon}
+        field={"code"}
       />
       <div className="flex gap-2">
-        <CouponSelect value={newCoupon.discountType} />
+        <CouponSelect
+          value={newCoupon.discountType}
+          updateNewCoupon={updateNewCoupon}
+        />
+        <CouponInput
+          type={"number"}
+          value={newCoupon.discountValue}
+          placeholder={"할인 값"}
+          updateNewCoupon={updateNewCoupon}
+          field={"discountValue"}
+        />
       </div>
-
       <button
-        onClick={() => handleAddCoupon()}
+        onClick={onAddCoupon}
         className="w-full bg-green-500 text-white p-2 rounded hover:bg-green-600"
       >
         쿠폰 추가
@@ -36,3 +47,5 @@ export const CouponForm = () => {
     </div>
   );
 };
+
+export default CouponForm;
