+++ src/refactoring/components/admin/DiscountForm.tsx
@@ -1,8 +1,8 @@
-interface DiscountFormProps {
-  newDiscount: { quantity: number; rate: number };
-  setNewDiscount: React.Dispatch<
-    React.SetStateAction<{ quantity: number; rate: number }>
-  >;
+import { Discount } from "@/types";
+
+interface Props {
+  newDiscount: Discount;
+  setNewDiscount: React.Dispatch<React.SetStateAction<Discount>>;
   onAddDiscount: () => void;
 }
 
@@ -10,39 +10,48 @@ const DiscountForm = ({
   newDiscount,
   setNewDiscount,
   onAddDiscount,
-}: DiscountFormProps) => {
+}: Props) => {
+  const handleQuantityChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setNewDiscount({
+      ...newDiscount,
+      quantity: parseInt(e.target.value),
+    });
+  };
+
+  const handleRateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setNewDiscount({
+      ...newDiscount,
+      rate: parseFloat(e.target.value) / 100,
+    });
+  };
+
   return (
-    <div className="flex space-x-2">
-      <input
-        type="number"
-        placeholder="수량"
-        value={newDiscount.quantity}
-        onChange={(e) =>
-          setNewDiscount({
-            ...newDiscount,
-            quantity: parseInt(e.target.value),
-          })
-        }
-        className="w-1/3 p-2 border rounded"
-      />
-      <input
-        type="number"
-        placeholder="할인율 (%)"
-        value={newDiscount.rate * 100}
-        onChange={(e) =>
-          setNewDiscount({
-            ...newDiscount,
-            rate: parseInt(e.target.value) / 100,
-          })
-        }
-        className="w-1/3 p-2 border rounded"
-      />
-      <button
-        onClick={onAddDiscount}
-        className="w-1/3 bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-      >
-        할인 추가
-      </button>
+    <div className="mt-2">
+      <div className="flex flex-wrap gap-2 mb-2">
+        <input
+          type="number"
+          placeholder="수량"
+          value={newDiscount.quantity}
+          onChange={handleQuantityChange}
+          className="flex-1 p-2 border rounded"
+          min="1"
+        />
+        <input
+          type="number"
+          placeholder="할인율 (%)"
+          value={newDiscount.rate * 100}
+          onChange={handleRateChange}
+          className="flex-1 p-2 border rounded"
+          min="0"
+          max="100"
+        />
+        <button
+          onClick={onAddDiscount}
+          className="bg-blue-500 text-white px-4 py-2 rounded hover:bg-blue-600"
+        >
+          할인 추가
+        </button>
+      </div>
     </div>
   );
 };
