+++ src/refactoring/components/admin/NewProductForm.tsx
@@ -1,4 +1,4 @@
-import { Product } from "../../../../types";
+import { Product } from "@/types";
 
 interface Props {
   newProduct: Omit<Product, "id">;
