+++ src/refactoring/components/admin/ProductEditForm.tsx
@@ -1,22 +1,40 @@
-import { useAtomValue, useSetAtom } from "jotai";
-import { formatDiscountInfo } from "../../models/product.ts";
-import { DiscountFormSection } from "./DiscountFormSection";
-import { editingProductAtom } from "../../store/products/atom.ts";
-import {
-  handleUpdateEditingFieldAtom,
-  handleCancelEditAtom,
-  handleRemoveDiscountAtom,
-  handleSaveProductAtom,
-} from "../../store/products/actions.ts";
+import { useState } from "react";
+import DiscountForm from "./DiscountForm";
+import { Discount, Product } from "@/types.ts";
+import { formatDiscountInfo } from "@/refactoring/models/product";
 
-export const ProductEditForm = () => {
-  const editingProduct = useAtomValue(editingProductAtom);
-  const handleUpdateEditingField = useSetAtom(handleUpdateEditingFieldAtom);
-  const handleCancelEdit = useSetAtom(handleCancelEditAtom);
-  const handleRemoveDiscount = useSetAtom(handleRemoveDiscountAtom);
-  const handleSaveProduct = useSetAtom(handleSaveProductAtom);
+interface Props {
+  editingProduct: Product;
+  handleUpdateEditingField: (field: keyof Product, value: any) => void;
+  onCancelEdit: () => void;
+  onSaveProduct: () => void;
+}
 
-  if (!editingProduct) return null;
+const ProductEditForm = ({
+  editingProduct,
+  handleUpdateEditingField,
+  onCancelEdit,
+  onSaveProduct,
+}: Props) => {
+  const [newDiscount, setNewDiscount] = useState<Discount>({
+    quantity: 0,
+    rate: 0,
+  });
+
+  const handleAddDiscount = () => {
+    handleUpdateEditingField("discounts", [
+      ...editingProduct.discounts,
+      newDiscount,
+    ]);
+    setNewDiscount({ quantity: 0, rate: 0 });
+  };
+
+  const handleRemoveDiscount = (index: number) => {
+    handleUpdateEditingField(
+      "discounts",
+      editingProduct.discounts.filter((_, i) => i !== index)
+    );
+  };
 
   return (
     <div>
@@ -25,12 +43,7 @@ export const ProductEditForm = () => {
         <input
           type="text"
           value={editingProduct.name}
-          onChange={(e) =>
-            handleUpdateEditingField({
-              field: "name",
-              value: e.target.value,
-            })
-          }
+          onChange={(e) => handleUpdateEditingField("name", e.target.value)}
           className="w-full p-2 border rounded"
         />
       </div>
@@ -40,10 +53,7 @@ export const ProductEditForm = () => {
           type="number"
           value={editingProduct.price}
           onChange={(e) =>
-            handleUpdateEditingField({
-              field: "price",
-              value: parseInt(e.target.value),
-            })
+            handleUpdateEditingField("price", parseInt(e.target.value))
           }
           className="w-full p-2 border rounded"
         />
@@ -54,10 +64,7 @@ export const ProductEditForm = () => {
           type="number"
           value={editingProduct.stock}
           onChange={(e) =>
-            handleUpdateEditingField({
-              field: "stock",
-              value: parseInt(e.target.value),
-            })
+            handleUpdateEditingField("stock", parseInt(e.target.value))
           }
           className="w-full p-2 border rounded"
         />
@@ -77,18 +84,22 @@ export const ProductEditForm = () => {
             </button>
           </div>
         ))}
-        <DiscountFormSection />
+        <DiscountForm
+          newDiscount={newDiscount}
+          setNewDiscount={setNewDiscount}
+          onAddDiscount={handleAddDiscount}
+        />
       </div>
 
       <div className="mt-2 space-x-2">
         <button
-          onClick={() => handleSaveProduct()}
+          onClick={onSaveProduct}
           className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600"
         >
           수정 완료
         </button>
         <button
-          onClick={() => handleCancelEdit()}
+          onClick={onCancelEdit}
           className="bg-gray-500 text-white px-2 py-1 rounded hover:bg-gray-600"
         >
           취소
@@ -97,3 +108,5 @@ export const ProductEditForm = () => {
     </div>
   );
 };
+
+export default ProductEditForm;
