+++ src/refactoring/components/afterRefactoring/CouponButton.tsx
