+++ src/refactoring/components/afterRefactoring/DiscountFormSection.tsx
