+++ src/refactoring/components/afterRefactoring/NewProductSection.tsx
