+++ src/refactoring/components/afterRefactoring/ProductDetail.tsx
