+++ src/refactoring/components/afterRefactoring/ProductEditForm.tsx
@@ -1,40 +1,22 @@
-import { useState } from "react";
-import { formatDiscountInfo } from "../../../models/product.ts";
-import DiscountForm from "./DiscountForm";
-import { Discount, Product } from "../../../../types.ts";
+import { useAtomValue, useSetAtom } from "jotai";
+import { formatDiscountInfo } from "../../models/product.ts";
+import { DiscountFormSection } from "./DiscountFormSection";
+import { editingProductAtom } from "../../store/products/atom.ts";
+import {
+  handleUpdateEditingFieldAtom,
+  handleCancelEditAtom,
+  handleRemoveDiscountAtom,
+  handleSaveProductAtom,
+} from "../../store/products/actions.ts";
 
-interface Props {
-  editingProduct: Product;
-  handleUpdateEditingField: (field: keyof Product, value: any) => void;
-  onCancelEdit: () => void;
-  onSaveProduct: () => void;
-}
+export const ProductEditForm = () => {
+  const editingProduct = useAtomValue(editingProductAtom);
+  const handleUpdateEditingField = useSetAtom(handleUpdateEditingFieldAtom);
+  const handleCancelEdit = useSetAtom(handleCancelEditAtom);
+  const handleRemoveDiscount = useSetAtom(handleRemoveDiscountAtom);
+  const handleSaveProduct = useSetAtom(handleSaveProductAtom);
 
-const ProductEditForm = ({
-  editingProduct,
-  handleUpdateEditingField,
-  onCancelEdit,
-  onSaveProduct,
-}: Props) => {
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0,
-  });
-
-  const handleAddDiscount = () => {
-    handleUpdateEditingField("discounts", [
-      ...editingProduct.discounts,
-      newDiscount,
-    ]);
-    setNewDiscount({ quantity: 0, rate: 0 });
-  };
-
-  const handleRemoveDiscount = (index: number) => {
-    handleUpdateEditingField(
-      "discounts",
-      editingProduct.discounts.filter((_, i) => i !== index)
-    );
-  };
+  if (!editingProduct) return null;
 
   return (
     <div>
@@ -43,7 +25,12 @@ const ProductEditForm = ({
         <input
           type="text"
           value={editingProduct.name}
-          onChange={(e) => handleUpdateEditingField("name", e.target.value)}
+          onChange={(e) =>
+            handleUpdateEditingField({
+              field: "name",
+              value: e.target.value,
+            })
+          }
           className="w-full p-2 border rounded"
         />
       </div>
@@ -53,7 +40,10 @@ const ProductEditForm = ({
           type="number"
           value={editingProduct.price}
           onChange={(e) =>
-            handleUpdateEditingField("price", parseInt(e.target.value))
+            handleUpdateEditingField({
+              field: "price",
+              value: parseInt(e.target.value),
+            })
           }
           className="w-full p-2 border rounded"
         />
@@ -64,7 +54,10 @@ const ProductEditForm = ({
           type="number"
           value={editingProduct.stock}
           onChange={(e) =>
-            handleUpdateEditingField("stock", parseInt(e.target.value))
+            handleUpdateEditingField({
+              field: "stock",
+              value: parseInt(e.target.value),
+            })
           }
           className="w-full p-2 border rounded"
         />
@@ -84,22 +77,18 @@ const ProductEditForm = ({
             </button>
           </div>
         ))}
-        <DiscountForm
-          newDiscount={newDiscount}
-          setNewDiscount={setNewDiscount}
-          onAddDiscount={handleAddDiscount}
-        />
+        <DiscountFormSection />
       </div>
 
       <div className="mt-2 space-x-2">
         <button
-          onClick={onSaveProduct}
+          onClick={() => handleSaveProduct()}
           className="bg-green-500 text-white px-2 py-1 rounded hover:bg-green-600"
         >
           수정 완료
         </button>
         <button
-          onClick={onCancelEdit}
+          onClick={() => handleCancelEdit()}
           className="bg-gray-500 text-white px-2 py-1 rounded hover:bg-gray-600"
         >
           취소
@@ -108,5 +97,3 @@ const ProductEditForm = ({
     </div>
   );
 };
-
-export default ProductEditForm;
