+++ src/refactoring/components/afterRefactoring/ProductManagement.tsx
@@ -1,0 +1,16 @@
+import { useAtomValue } from "jotai";
+import { showNewProductFormAtom } from "../../store/products/atom.ts";
+import { ProductList } from "./ProductList.tsx";
+import { NewProductSection } from "./NewProductSection.tsx";
+
+export const ProductManagement = () => {
+  const showNewProductForm = useAtomValue(showNewProductFormAtom);
+
+  return (
+    <div>
+      <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
+      <NewProductSection />
+      {!showNewProductForm && <ProductList />}
+    </div>
+  );
+};
